+++ src/utils/rxjs.ts
@@ -1,5 +1,5 @@
 import { Observable, of } from "rxjs";
 
-export function emptyNext(): Observable<undefined> {
+export function emptyNext(): Observable<void> {
 	return of(undefined);
 }