+++ CHANGELOG.md
@@ -2,9 +2,12 @@
 
 ### Features
 
-**hub:**
-- connect now will throw error after being exhausted
-- 
+- **hub connection:** connect now will throw error after being exhausted
+
+### Bug Fixes
+
+- **hub connection:** when connecting and fails, during retring if `disconnect` is triggered was being ignored and will keep connecting
+
 
 <a name="1.0.1"></a>
 # [1.0.1](https://github.com/sketch7/signalr-client/compare/1.0.0...1.0.1) (2019-02-05)